# apps/backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /users:
    post:
      operationId: UsersController_create
      parameters:
        []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
      tags:
        - Users
    get:
      operationId: UsersController_findAll
      parameters:
        []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponseDto"
      tags:
        - Users
...
  "/users/{id}":
    get:
      operationId: UsersController_findOne
      parameters:
        -
          name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
      tags:
        - Users
    patch:
      operationId: UsersController_update
      parameters:
        -
          name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
      tags:
        - Users
    delete:
      operationId: UsersController_remove
      parameters:
        -
          name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: string
      tags:
        - Users
  /users/avatar:
    post:
      operationId: UsersController_addAvatar
      parameters:
        []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarResponseDto"
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarResponseDto"
      tags:
        - Users
  "/users/avatar/{userId}":
    delete:
      operationId: UsersController_deleteAvatar
      parameters:
        -
          name: userId
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponseDto"
      tags:
        - Users
...
  /profiles/user:
    put:
      operationId: ProfilesController_updateUserProfile
      parameters:
        []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserProfileDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileResponseDto"
      tags:
        - Profiles
    get:
      operationId: ProfilesController_getUserProfile
      parameters:
        []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithProfileResponseDto"
      tags:
        - Profiles
  /profiles/doc:
    put:
      operationId: ProfilesController_updateDocProfile
      parameters:
        []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocProfileDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocProfileResponseDto"
      tags:
        - Profiles
    get:
      operationId: ProfilesController_getDocProfile
      parameters:
        []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocWithProfileResponseDto"
      tags:
        - Profiles
...
  /profiles/admin:
    post:
      operationId: ProfilesController_updateAdminProfile
      parameters:
        []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminProfileDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProfileResponseDto"
      tags:
        - Profiles
  /profiles/rate:
    post:
      operationId: ProfilesController_addDocRate
      parameters:
        []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocRateDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocProfileResponseDto"
      tags:
        - Profiles
  "/profiles/doc/{docId}/ratings":
    get:
      operationId: ProfilesController_getDocRatings
      parameters:
        -
          name: docId
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocRatingsResponseDto"
      tags:
        - Profiles
...
  /payments/checkout:
    post:
      operationId: PaymentsController_createCheckoutSession
      parameters:
        []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutSessionDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponseDto"
      tags:
        - Payments
  /payments/p24/start:
    post:
      operationId: PaymentsController_startP24
      parameters:
        []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutSessionDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/P24StartResponseDto"
      tags:
        - Payments
...
info:
  title: Melius API
  description: Melius backend API
  version: "1.0"
  contact:
    {}
...
servers:
  []
components:
  schemas:
...
    Role:
      type: string
      enum:
        - user
        - doc
        - admin
...
    OkResponseDto:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
    Language:
      type: string
      enum:
        - pl
        - en
    CreateUserDto:
      type: object
      properties:
        role:
          allOf:
            -
              $ref: "#/components/schemas/Role"
        language:
          allOf:
            -
              $ref: "#/components/schemas/Language"
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        consentTerms:
          type: boolean
        consentAdult:
          type: boolean
        consentHealthData:
          type: boolean
      required:
        - email
        - password
        - consentTerms
        - consentAdult
        - consentHealthData
    UserResponseDto:
      type: object
      properties:
        id:
          type: number
        role:
          allOf:
            -
              $ref: "#/components/schemas/Role"
        email:
          type: string
      required:
        - id
        - role
        - email
    UpdateUserDto:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required:
        - email
        - password
    AvatarResponseDto:
      type: object
      properties:
        url:
          type: string
      required:
        - url
...
    UpdateUserProfileDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        problems:
          minItems: 1
          type: array
          items:
            type: number
      required:
        - firstName
        - lastName
        - problems
    ProblemEntity:
      type: object
      properties:
        id:
          type: number
        problemKey:
          type: string
      required:
        - id
        - problemKey
    UserProfileResponseDto:
      type: object
      properties:
        userId:
          type: number
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        problems:
          type: array
          items:
            $ref: "#/components/schemas/ProblemEntity"
      required:
        - userId
        - firstName
        - lastName
        - problems
    Profession:
      type: string
      enum:
        - psychologist
        - psychotherapist
        - sexologist
    UpdateDocProfileDto:
      type: object
      properties:
        profession:
          allOf:
            -
              $ref: "#/components/schemas/Profession"
        firstName:
          type: string
        lastName:
          type: string
        sessionPricePln:
          type: number
          minimum: 50
          maximum: 1000
        specializations:
          minItems: 1
          type: array
          items:
            type: number
        docTermsAccepted:
          type: boolean
      required:
        - profession
        - firstName
        - lastName
        - specializations
    SpecializationEntity:
      type: object
      properties:
        id:
          type: number
        specializationKey:
          type: string
      required:
        - id
        - specializationKey
    DocVerificationStatus:
      type: string
      enum:
        - draft
        - submitted
        - approved
        - rejected
    DocProfileResponseDto:
      type: object
      properties:
        docId:
          type: number
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        profession:
          nullable: true
          allOf:
            -
              $ref: "#/components/schemas/Profession"
        rate:
          type: number
          nullable: true
        ratesLot:
          type: number
          nullable: true
        sessionPricePln:
          type: number
          nullable: true
        specializations:
          type: array
          items:
            $ref: "#/components/schemas/SpecializationEntity"
        verificationStatus:
          allOf:
            -
              $ref: "#/components/schemas/DocVerificationStatus"
        rejectionReason:
          type: string
          nullable: true
      required:
        - docId
        - firstName
        - lastName
        - profession
        - rate
        - ratesLot
        - sessionPricePln
        - specializations
        - verificationStatus
        - rejectionReason
    UpdateAdminProfileDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
      required:
        - firstName
        - lastName
    AdminProfileResponseDto:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        createdAt:
          format: "date-time"
          type: string
        updatedAt:
          format: "date-time"
          type: string
      required:
        - id
        - userId
        - firstName
        - lastName
        - createdAt
        - updatedAt
    UserWithProfileResponseDto:
      type: object
      properties:
        id:
          type: number
        role:
          allOf:
            -
              $ref: "#/components/schemas/Role"
        email:
          type: string
        avatar:
          nullable: true
          allOf:
            -
              $ref: "#/components/schemas/AvatarResponseDto"
        userProfile:
          $ref: "#/components/schemas/UserProfileResponseDto"
      required:
        - id
        - role
        - email
        - avatar
        - userProfile
    DocWithProfileResponseDto:
      type: object
      properties:
        id:
          type: number
        role:
          allOf:
            -
              $ref: "#/components/schemas/Role"
        email:
          type: string
        avatar:
          nullable: true
          allOf:
            -
              $ref: "#/components/schemas/AvatarResponseDto"
        docProfile:
          $ref: "#/components/schemas/DocProfileResponseDto"
      required:
        - id
        - role
        - email
        - avatar
        - docProfile
    CreateDocRateDto:
      type: object
      properties:
        docId:
          type: number
          minimum: 1
        meetingId:
          type: number
          minimum: 1
        rate:
          type: number
          minimum: 1
          maximum: 5
        comment:
          type: string
          maxLength: 1000
      required:
        - docId
        - meetingId
        - rate
    DocRatingUserDto:
      type: object
      properties:
        id:
          type: number
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
      required:
        - id
        - firstName
        - lastName
    DocRatingItemDto:
      type: object
      properties:
        id:
          type: number
        rate:
          type: number
        comment:
          type: string
          nullable: true
        createdAt:
          format: "date-time"
          type: string
        user:
          $ref: "#/components/schemas/DocRatingUserDto"
      required:
        - id
        - rate
        - comment
        - createdAt
        - user
    DocRatingsResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DocRatingItemDto"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - items
        - total
        - page
        - limit
...
    CreateCheckoutSessionDto:
      type: object
      properties:
        meetingId:
          type: number
      required:
        - meetingId
    CheckoutSessionResponseDto:
      type: object
      properties:
        url:
          type: string
        sessionId:
          type: string
      required:
        - url
        - sessionId
    P24StartResponseDto:
      type: object
      properties:
        url:
          type: string
        token:
          type: string
        sessionId:
          type: string
      required:
        - url
        - token
        - sessionId
...